Question Answering System/rag-based-question-answering-system/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.2.1
numpy==1.26.4
pypdf==5.0.1
python-dotenv==1.0.1
openai==1.51.2
httpx==0.27.2